test: . Move functions for verifying statuses to TestingTools/VerifyStatuses.ts
